remove /proc based IO accounting
The /proc based IO accounting tries to monitor the amount of data
written to the data storage. This is a nice idea, but seems not
to work very well in practise due to Linux caching a lot of writes
first (leads to a jumpy progress bar).

This approach likely also does not work well for compressed file
systems (UBIFS...).

The new approach uses the utility pv (pipe viewer) which monitors
the effective, uncompressed data stream.