add module clean function
Add module clean function using blkdiscard. What is a bit unfortunate
is that we can't clean the U-Boot environment: The blkdiscard deletes
in erase_size steps, which is in current Toshiba eMMC 512kB. Since the
U-Boot environment is just one block before the config block, we can't
delete the U-Boot environment using blkdiscard.