readd line if we tried to parse progress from it
We use stdout to read progress of flash commands using pv (pipe
viewer). This works nicely, however in the error case we try to
parse a line and opt out in case it is not a number. This first
line is then typically missing from the error message. Readd the
line we tried to parse in case an error happens.