use part of URL to store temporary files
Using the image name is not safe, as we don't have the module name
in the image name (e.g. Colibri/Apalis images have the same name).
Also make sure that we can distingush two images even if they
generate existing temporary folders.

Note: We could use a safe/on-board mechanism to generate temporary
directories, but for debugability it is nicer to have names which
mean something.