unify content class, remove/rename DTO classes
The exact behavior and valid properties for content of different
mass storage classes differs (e.g. MTD vs. block device). We could
make use of class inheritance to properly reflect all that. But
for the sake of simplicity we avoid that for now, and use a single
ContentInfo class for block devices, MTD devices and UBI volumes.
The implementation however is type specific, and uses the properties
according to the type of mass storage device.

Signed-off-by: Stefan Agner <stefan@agner.ch>