update UBI volumes over internet too
Use wget and stdin to update UBI volumes directly from the network.
Note that this is somewhat more complex: ubiupdatevol requires the
size to flash in advance... We use wget to get only the header and
use Content-Size to figure out the exact file size. Note that this
needs Server support, but it is the only way we currently can handle
that case.

We could add a JSON property which has the exact size, but this is
rather error prone (e.g. when somebody replaces the device tree
without updating the image.json file...)

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>