fix NAND flash from Internet
Flashing to a UBI volume directly from Internet requires
to get size of the file first. This is done using a wget
command.

Since commit ec45481a8894 ("avoid dupplication of MTD erase
function") commands get now executed in the root folder if
not specified otherwise.

The parameters used in the wget command to get the http
headers turn out to actually write something to the current
folder. Since the root folder is read-only this writing
failed and wget exited with error code 3 (File I/O error).
This lead Tezi to print an error message "Failed to get
size of URL"...

Rectify this by using wget parameters which do not attempt
to write the content of the file. Also improve the error
message.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>