fail gracefully when file does not exist
If a file to be untared/copied does not exist, the pipe is not
executed completely. In that case, the Qt UI tries to read from
the pipeviewer FIFO. However, the write side of the FIFO is never
opened, which leads to the read side to block forever.

Unfortunately it is not easy to detect whether a read side exists.
It seems that normally the pipe viewer thread is already finished
when the process ends. If not, make sure that somebody opened the
write side at least once by simply writing a newline into the
pipe.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>