fix dynamic resizing
In some cases calculation of the text height from the Description + URI
in the image items is not accurate because the rectangles used are not
the same dimensions than those used during drawing. This causes text on
selected images to overflow when the URI text wraps into a newline.
Also, it is a better idea to set the Description and URI fields in
separate roles and combine them together when adequate than to deal with
them in a block. This also fixes the newline that was lacking in the
tooltips.

Signed-off-by: Gerard Salvatella <gerard.salvatella@toradex.com>