qt5 port: adapt json parser
The json facilities have been changed in Qt5. Adapt the json class to be
able to use the json parser. Note that the valijson hasn't been modified
as it still works as-is with the rapidjson parser as binding, although
a direct, more convenient binding with QJson should be devised in the
future.

Signed-off-by: Gerard Salvatella <gerard.salvatella@toradex.com>