use image index to supply the correct icon
Currently downloaded icons from images in a server are assigned to the
images based on the icon filename. This causes icons with the same
filename (but otherwise different) to not be properly assigned to the
correct image. Use the index of the image to assign the corresponding
icon to it.

Signed-off-by: Gerard Salvatella <gerard.salvatella@toradex.com>