avoid reading rpmb partition on emmcs
Attempting to read an rpmb partition in emmcs on imx8 produces an i/o
error in kernel that hangs it for some time. Tezi does check all
partitions in an emmc to look for images, but the above problem makes it
hang for a while during an installation. Avoiding lookup on this
partition solves the problem.

Signed-off-by: Gerard Salvatella <gerard.salvatella@toradex.com>