main.cpp: store system clock into /etc/timestamp
We now have a process running at system starting up, to check the
timestamp in /etc/timestamp, in order to set system time according to
the value read from it, in case it's more recent than the current time.

We hence also need to update the /etc/timestamp with the current time
while system is shutting down.

Signed-off-by: Ming Liu <ming.liu@toradex.com>