remove language files
We did not translate Tezi so far. The language files at this point are
very much outdated and probably not really valueable anymore. Currently
they are very annoying because lots of (automatic) changes are in them
which pollute git commits. Just delete them, we could still restore
them in case we add translation to Tezi.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>