multiimagewritethread: fix deadlock when reading from progress pipe
Sometimes during image installation the installation screen is stuck
(often showing no progress). The tezi.log contains the following
message:
  Pipeviewer poll did not finish on its own, make sure pipe got opened and received data.

It seems that a race between the FIFO reader thread and the actual
pipe viewer process (pv) causes Qt Tezi to deadlock.

It showed that the deadlock could be unlocked by manually read from the
pipe, hence the problem must be on the write side (the actual work
around seems to cause the deadlock).

Fix the issue in two ways:
- Only apply the work around when really necessary (when the pipe
  actually knowingly failed) by moving it into the exit code if
  statement.
- Write to the pipe using non-blocking mode.

Related-to: TEI-28
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>