Enable RNDIS device before reading configblock
Currently we have a problem that when reading configblock fails, RNDIS
device would not be enabled, in which case, a end user could not access
Tezi by VNC through RNDIS network.

To fix that, we need enable RNDIS device before reading configblock,
but at that moment, we dont have serial number, product id and so on
information, so we need initRndis and enableRndis to be re-enterable so
they could be called before and after reading configblock.

To achieve that, we introduce two functions setMassStorageAttrs and
setRndisAttrs, that they could be called to set serial number, product
id, and so on attributes and make usbgadget_ms_init and
usbgadget_rndis_init re-enterable by moving in descriptor deleting
parts.

Related-to: TEI-539
Signed-off-by: Ming Liu <ming.liu@toradex.com>