config.h: bump IMAGE_CONFIG_FORMAT to 4
For i.MX8 processors there was a bug in Tezi since the port of tezi to
i.MX8. Early i.MX8 Tezi versions had a default u-boot environment
compiled in the command fw_setenv that is used to write u-boot
environment. This lead to newer images not booting.

To distinguish Tezi version that have this behavior fixed this commit
bumps the image config format so we can enforce a minimal Tezi version
in our Images built.

Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>