imagelist: fix a autoinstall hang issue
When 'autoinstall' entry is enabled in a image.json file, it is
observed Tezi hangs when installing that image. This is due to a
deadlock in addImages function, listMutex must be released before
emitting foundAutoInstallImage, or else it would try to get it again
in the slot function.

Also added a downloadAborted slot in ImageListDownload class, so it
could be aborted when abortAllDownloads is emitted.

Related-to: TEI-773

Signed-off-by: Ming Liu <ming.liu@toradex.com>