imagelist: Fix wrong comparison of tezi versions
We use different version strings (DISTRO_VERSION) depending on the
release type, e.g.:

Quarterly:	5.3.0+build.10
Monthly:	5.5.0-devel-202111+build.15
Nightly:	5.5.0-devel-20211109+build.480
Manual-Build:	5.5.0-devel-20211109123456+build.0

However the versioning we pass to U-Boot/Kernel and Qt-Tezi is a
reduced form (TDX_VERSION). Here we only distinguish between e.g.:

Release:	5.5.0
Prerelease:	5.5.0-devel

The existing code does obviously not take this fact into account, as it
compares TDX_VERSION stored in qt-tezi with DISTRO_VERSION that is
stored in image.json of the image that is parsed.

Fix this by only checking if there is a "devel" string behind the
version number or not.

This commit has the downside that within prereleases (Monthly and
Nightly) one is not able to autoinstall a Tezi.

It however fixes a behaviour where Tezi goes into an install-loop of
continuously installing the same version again and again, since it
used to compare String length of a tezi version (TDX_VERSION)
of e.g. (5.5.0-devel) to a version in a image.json (DISTRO_VERSION) of
e.g. (5.5.0-devel-202111+build.15). And obviously the later one is
always longer.

Related-to: TEI-964
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>