change message if image is to new for this tezi
User validation has shown, that the current message may not be
interpreted correctly. This message is shown if the Config Format of the
image to be installed is newer than what the current tezi version can do
hence the user should update tezi first before installing the desired
image.

Change this message for better clarity.

Related-to: TEI-968
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>