mainwindow: add images only to list if necessary
If there are no images in ui->list downloadImage should not add any
image to this list. downloadImage is meant to directly install an image
from tezictl.

What happened before was, that addImages was called which further down
the chain changed installationStatus to idle because there where still
no images in the list which lead to another function change the status
to SCANNNING and subsequently back to IDLE which was wrong.

Since it does not make sense beforehand to add any images to a list
which anyway get installed remove this connection.

Related-to: TEI-1016
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>