configblock: add assembly revision 26-99
So far we used V1.1A to V1.1Z [A-Z]. However in theory we have a range
assigned from 0-99. Display versions > 25 in plain numbers after a
number sign (e.g. V1.1#26).

So far a user was only allowed to enter a valid version number. This is
no longer possible and loosened by this commit. A user could potentially
enter a version V1.2355 which ends up as V1.2#99 as the number "3"
should be a capital letter or "#".

Hence this commit introduces version 99 to be an error-case.

Related-to: TEI-1017, TEI-1018
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>