multiimagewritethread: set boot partition 0 to active
If the module was previously switching to boot from the second eMMC boot
partition because of an update (e.g. Torizon U-Boot update) we would not
boot the right U-Boot after we install a new image. This commit changes
that by checking if we update the bootloader and if so we will switch to
the first boot partition if the second one should currently be active.

Related-to: TEI-1035
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>