usbgadgethelper.c: move to setup a ncm gadget, rather than rndis
Follow the changes as done with the config schema in the BSP.

For a Windows host setting WINNCM as the compatible ID, the UsbNcm
driver is loaded on enumeration without the need for a custom driver
or inf file. [1].

Setting the function attrs explicitely returns an error from libusbgx.
The library sets the (RNDIS) attrs class/subclass/protocol which do
not exist for the NCM gadget and thus fails. As we do anyway
set all attrs to their defaults just drop setting them.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/drivers/usb/gadget/function/f_ncm.c?h=v6.3&id=793409292382027226769d0299987f06cbd97a6e

Related-to: ELB-1061
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>