httpapi: Fix image index out-of-bounds
If one triggers the install image endpoint with an image index
(`imageId`) that is out-of-bounds even though qt-tezi sets `status` to
`BAD_REQUEST`, it currently proceeds to the installation, which causes
qt-tezi to close suddenly.

This commit fixes this issue by adding a `break` after the `BAD_REQUEST`,
so the bad `imageId` won't be used for installation.

Signed-off-by: Henrique Garcia <henrique.garcia@toradex.com>
(cherry picked from commit b92d518a0c5447e21693c232fbce7cfcac608c91)
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>