gitlab-ci: serialize jobs and keep the tezi binary
Hopefully this helps to prevent the following ci error when both jobs
try to fetch at the same time and one of them fails with:

| Fetching changes...
| Initialized empty Git repository in /builds/max.krummenacher/qt-tezi/.git/
| Created fresh repository.
| Checking out 4901a692 as master...
| fatal: Unable to create '/builds/max.krummenacher/qt-tezi/.git/index.lock': File exists.
| Another git process seems to be running in this repository, e.g.
| ...

Additionally it allows to get the tezi binary from the CI build. So
one can try it out by copying it to /tmp on a running tezi.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>