Use U-Boot environment as default
Currently, the U-Boot environment passed on "u-boot_env" adds or
overwrites variables of an existing environment. This happens because
fw_setenv needs a default environment else it fails.

When fw_setenv starts, it tries to load the environment from the device
described on the config file, otherwise, it uses a default environment
file (by default: /etc/u-boot-initial-env)).
As a consequence, the final environment may be different from the one
delivered on the image.

Change fw_setenv command to use the "u-boot-env" property argument as
the default and final environment.

Related-to: ELB-5467, TEI-1020

Signed-off-by: Vitor Soares <vitor.soares@toradex.com>