Change fw_setenv call to use /dev/null as default
Currently, we are passing the u-boot-initial-env from the image as
the default environment, although it is not writing to the memory.

Tezi 5 uses an older version of libubootenv (0.3.1), which doesn't
save the default environment in the flash memory when the script
argument is NULL.

Fix this by changing the fw_setenv command call to pass /dev/null as
the default environment (to not fall back in /etc/u-boot-initial-env)
and the target image u-boot-initial-env as the script argument.

Related-to: TEI-1020, ELB-5467 and ELB-4940
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>