multiimagewritethread: add input parameter to runCommand method
runCommand method is used to run an executable and waiting its execution.
It was not possible to provide a standard input to the command.
So, add an input parameter which contains data to be provided as standard
input.

Related-to: TEI-1178
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>
(cherry picked from commit d9d74330c86a5fcd946caf1d1f5c91b98d5ee5a1)