use generic linux guid partition type for gpt and raw
The gpt GUID 00000000-0000-0000-0000-000000000000 set for 'raw' gpt
partitions by qt-tezi is considered an unused entry in a gpt partition
table.
At least with a current kernel / udev no entry for the block device
is created in /dev/. Due to that waiting for all greated partitions
to appear in /dev/ fails with a timeout.
qt-tezi thus fails with
`Error: "Timeout waiting for partition device /dev/mmcblk2p3."` when
creating a raw partition with gpt partition table if image.json
contains such a raw partition.

Fix that by using the generic Linux filesystem data GUID

Compare with:
https://en.wikipedia.org/wiki/GUID_Partition_Table#Partition_type_GUIDs
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/block/partitions/efi.c?h=v6.1#n454

Related-to: TEI-1167
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>