multiimagewritethread: fix last partition end sector calculation
When there are one or more maximized partitions, space is evenly distributed
among the partitions.
In some cases, particularly if the first partition offset is bigger then 4MB,
the calculated end sector of the last partition exceeds the last device sector.
The current version of sfdisk fails to create such partitions whereas
the older version adjusts the last partition to fit within the final sector.
Fix the last partition end sector calculation when the partition must be maximized.

Related-to: TEI-1197
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>