multiimagewritethread: reduce config block restore time
The boot partition is normally marked for erasure, and the config block
is restored at the end of the full image programming. Typically, this time
frame is relatively short because the boot partition is
the last one to be updated.
However, since there are a some operations that may be lengthy and can be
forcibly interrupted, the config block might not be restored.
Reduce the risk of losing the config block by writing it as soon as
the partition is erased.

Related-to: TEI-1199
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>