mainwindow: wait for UDC to come up to initialize USB gadget
The UDC (USB Device Controller) may not be immediately available during
the initialization of the USB gadget. Specifically, NCM will not work in
this case as that gadget config is enabled in the main window init.

This happens when libusbgx is initialized before the UDC is ready. It
will not find UDCs added later as this is an atypical situation.

Poll for the UDC to appear before initializing any USB gadgets. This is
done with a new separate thread, UdcPollThread, so no other parts of the
application are slowed down.

Related-to: TEI-1300
Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>