mainwindow: fix use of _moduleInformation before initialization
The new USB gadget initialization flow introduced a bug by accessing
_moduleInformation->moduleSupported() before _moduleInformation was
initialized. On BSP7/Scarthgap this ran correctly by chance, but of
course the logic is wrong and the invalid access can also lead to
crashes, which is the case for BSP6/Kirkstone.

Fix this by respecting the correct order of operations.

Related-to: TEI-1300
Fixes: 87dbe839422e ("mainwindow: wait for UDC to come up to initialize USB gadget")
Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>